test: do not attempt drop if db is not created

+++ testing/src/main/kotlin/org/neo4j/connectors/kafka/testing/source/Neo4jSourceExtension.kt
@@ -174,10 +174,13 @@ internal class Neo4jSourceExtension(
       )
     }
 
+    @OptIn(ExperimentalAtomicApi::class)
     override fun close() {
       source?.unregister()
       if (!testFailed) {
-        driver?.dropDatabase(neo4jDatabase)
+        if (dbCreated.load()) {
+          driver?.dropDatabase(neo4jDatabase)
+        }
       }
       driver?.close()
